WW-3714 Move new Result class into result package
