+++ core/src/main/java/org/apache/struts2/ActionChainResult.java
@@ -18,10 +18,11 @@
  */
 package org.apache.struts2;
 
-import org.apache.struts2.inject.Inject;
-import org.apache.struts2.util.TextParseUtil;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.apache.struts2.inject.Inject;
+import org.apache.struts2.result.Result;
+import org.apache.struts2.util.TextParseUtil;
 
 import java.util.HashSet;
 import java.util.LinkedList;
