+++ core/src/main/java/org/apache/struts2/UnknownHandler.java
@@ -19,6 +19,7 @@
 package org.apache.struts2;
 
 import org.apache.struts2.config.entities.ActionConfig;
+import org.apache.struts2.result.Result;
 
 /**
  * Handles cases when the result or action is unknown.
