+++ core/src/main/java/org/apache/struts2/dispatcher/Dispatcher.java
@@ -18,36 +18,6 @@
  */
 package org.apache.struts2.dispatcher;
 
-import org.apache.struts2.ActionContext;
-import org.apache.struts2.ActionInvocation;
-import org.apache.struts2.ActionProxy;
-import org.apache.struts2.ActionProxyFactory;
-import org.apache.struts2.FileManager;
-import org.apache.struts2.FileManagerFactory;
-import org.apache.struts2.LocaleProviderFactory;
-import org.apache.struts2.ObjectFactory;
-import org.apache.struts2.Result;
-import org.apache.struts2.config.Configuration;
-import org.apache.struts2.config.ConfigurationException;
-import org.apache.struts2.config.ConfigurationManager;
-import org.apache.struts2.config.ConfigurationProvider;
-import org.apache.struts2.config.FileManagerFactoryProvider;
-import org.apache.struts2.config.FileManagerProvider;
-import org.apache.struts2.config.ServletContextAwareConfigurationProvider;
-import org.apache.struts2.config.entities.InterceptorMapping;
-import org.apache.struts2.config.entities.InterceptorStackConfig;
-import org.apache.struts2.config.entities.PackageConfig;
-import org.apache.struts2.config.providers.XmlConfigurationProvider;
-import org.apache.struts2.inject.Container;
-import org.apache.struts2.inject.ContainerBuilder;
-import org.apache.struts2.inject.Inject;
-import org.apache.struts2.interceptor.Interceptor;
-import org.apache.struts2.util.ClassLoaderUtil;
-import org.apache.struts2.util.ValueStack;
-import org.apache.struts2.util.ValueStackFactory;
-import org.apache.struts2.util.location.LocatableProperties;
-import org.apache.struts2.util.location.Location;
-import org.apache.struts2.util.location.LocationUtils;
 import jakarta.servlet.ServletContext;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
@@ -58,23 +28,53 @@ import org.apache.commons.lang3.StringUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.apache.logging.log4j.message.ParameterizedMessage;
+import org.apache.struts2.ActionContext;
+import org.apache.struts2.ActionInvocation;
+import org.apache.struts2.ActionProxy;
+import org.apache.struts2.ActionProxyFactory;
+import org.apache.struts2.FileManager;
+import org.apache.struts2.FileManagerFactory;
+import org.apache.struts2.LocaleProviderFactory;
+import org.apache.struts2.ObjectFactory;
 import org.apache.struts2.ServletActionContext;
 import org.apache.struts2.StrutsConstants;
 import org.apache.struts2.StrutsException;
 import org.apache.struts2.StrutsStatics;
+import org.apache.struts2.config.Configuration;
+import org.apache.struts2.config.ConfigurationException;
+import org.apache.struts2.config.ConfigurationManager;
+import org.apache.struts2.config.ConfigurationProvider;
 import org.apache.struts2.config.DefaultPropertiesProvider;
+import org.apache.struts2.config.FileManagerFactoryProvider;
+import org.apache.struts2.config.FileManagerProvider;
 import org.apache.struts2.config.PropertiesConfigurationProvider;
+import org.apache.struts2.config.ServletContextAwareConfigurationProvider;
 import org.apache.struts2.config.StrutsBeanSelectionProvider;
 import org.apache.struts2.config.StrutsJavaConfiguration;
 import org.apache.struts2.config.StrutsJavaConfigurationProvider;
 import org.apache.struts2.config.StrutsXmlConfigurationProvider;
+import org.apache.struts2.config.entities.InterceptorMapping;
+import org.apache.struts2.config.entities.InterceptorStackConfig;
+import org.apache.struts2.config.entities.PackageConfig;
+import org.apache.struts2.config.providers.XmlConfigurationProvider;
 import org.apache.struts2.dispatcher.mapper.ActionMapper;
 import org.apache.struts2.dispatcher.mapper.ActionMapping;
 import org.apache.struts2.dispatcher.multipart.MultiPartRequest;
 import org.apache.struts2.dispatcher.multipart.MultiPartRequestWrapper;
+import org.apache.struts2.inject.Container;
+import org.apache.struts2.inject.ContainerBuilder;
+import org.apache.struts2.inject.Inject;
+import org.apache.struts2.interceptor.Interceptor;
 import org.apache.struts2.ognl.ThreadAllowlist;
+import org.apache.struts2.result.Result;
+import org.apache.struts2.util.ClassLoaderUtil;
 import org.apache.struts2.util.ObjectFactoryDestroyable;
+import org.apache.struts2.util.ValueStack;
+import org.apache.struts2.util.ValueStackFactory;
 import org.apache.struts2.util.fs.JBossFileManager;
+import org.apache.struts2.util.location.LocatableProperties;
+import org.apache.struts2.util.location.Location;
+import org.apache.struts2.util.location.LocationUtils;
 
 import java.io.File;
 import java.io.IOException;
