+++ core/src/main/java/org/apache/struts2/interceptor/ChainingInterceptor.java
@@ -22,10 +22,10 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.apache.struts2.ActionChainResult;
 import org.apache.struts2.ActionInvocation;
-import org.apache.struts2.Result;
 import org.apache.struts2.StrutsConstants;
 import org.apache.struts2.Unchainable;
 import org.apache.struts2.inject.Inject;
+import org.apache.struts2.result.Result;
 import org.apache.struts2.util.CompoundRoot;
 import org.apache.struts2.util.ProxyUtil;
 import org.apache.struts2.util.TextParseUtil;
