+++ core/src/main/java/org/apache/struts2/interceptor/PreResultListener.java
@@ -19,10 +19,11 @@
 package org.apache.struts2.interceptor;
 
 import org.apache.struts2.ActionInvocation;
+import org.apache.struts2.result.Result;
 
 /**
  * PreResultListeners may be registered with an {@link ActionInvocation} to get a callback after the
- * {@link org.apache.struts2.Action} has been executed but before the {@link org.apache.struts2.Result}
+ * {@link org.apache.struts2.Action} has been executed but before the {@link Result}
  * is executed.
  *
  * @author Jason Carreira
@@ -31,7 +32,7 @@ public interface PreResultListener {
 
     /**
      * This callback method will be called after the {@link org.apache.struts2.Action} execution and
-     * before the {@link org.apache.struts2.Result} execution.
+     * before the {@link Result} execution.
      *
      * @param invocation  the action invocation
      * @param resultCode  the result code returned by the action (eg. <code>success</code>).
