+++ core/src/main/java/org/apache/struts2/mock/MockActionInvocation.java
@@ -22,8 +22,8 @@ import org.apache.struts2.ActionContext;
 import org.apache.struts2.ActionEventListener;
 import org.apache.struts2.ActionInvocation;
 import org.apache.struts2.ActionProxy;
-import org.apache.struts2.Result;
 import org.apache.struts2.interceptor.PreResultListener;
+import org.apache.struts2.result.Result;
 import org.apache.struts2.util.ValueStack;
 
 import java.util.ArrayList;
