+++ core/src/main/java/org/apache/struts2/mock/MockResult.java
@@ -19,7 +19,7 @@
 package org.apache.struts2.mock;
 
 import org.apache.struts2.ActionInvocation;
-import org.apache.struts2.Result;
+import org.apache.struts2.result.Result;
 
 /**
  * Mock for a {@link Result}.
