+++ plugins/convention/src/test/java/org/apache/struts2/convention/PackageBasedActionConfigBuilderTest.java
@@ -18,12 +18,14 @@
  */
 package org.apache.struts2.convention;
 
+import jakarta.servlet.ServletContext;
+import junit.framework.TestCase;
+import org.apache.commons.lang3.StringUtils;
 import org.apache.struts2.ActionChainResult;
 import org.apache.struts2.ActionContext;
 import org.apache.struts2.FileManager;
 import org.apache.struts2.FileManagerFactory;
 import org.apache.struts2.ObjectFactory;
-import org.apache.struts2.Result;
 import org.apache.struts2.config.Configuration;
 import org.apache.struts2.config.entities.ActionConfig;
 import org.apache.struts2.config.entities.ExceptionMappingConfig;
@@ -34,16 +36,6 @@ import org.apache.struts2.config.entities.PackageConfig;
 import org.apache.struts2.config.entities.ResultConfig;
 import org.apache.struts2.config.entities.ResultTypeConfig;
 import org.apache.struts2.config.impl.DefaultConfiguration;
-import org.apache.struts2.factory.DefaultInterceptorFactory;
-import org.apache.struts2.inject.Container;
-import org.apache.struts2.inject.Scope.Strategy;
-import org.apache.struts2.ognl.OgnlReflectionProvider;
-import org.apache.struts2.util.TextParseUtil;
-import org.apache.struts2.util.fs.DefaultFileManager;
-import org.apache.struts2.util.fs.DefaultFileManagerFactory;
-import org.apache.struts2.util.reflection.ReflectionException;
-import junit.framework.TestCase;
-import org.apache.commons.lang3.StringUtils;
 import org.apache.struts2.convention.actions.DefaultResultPathAction;
 import org.apache.struts2.convention.actions.NoAnnotationAction;
 import org.apache.struts2.convention.actions.Skip;
@@ -96,12 +88,20 @@ import org.apache.struts2.convention.actions.transactions.TransNameAction;
 import org.apache.struts2.convention.annotation.Action;
 import org.apache.struts2.convention.annotation.Actions;
 import org.apache.struts2.convention.dontfind.DontFindMeAction;
+import org.apache.struts2.factory.DefaultInterceptorFactory;
 import org.apache.struts2.factory.StrutsResultFactory;
+import org.apache.struts2.inject.Container;
+import org.apache.struts2.inject.Scope.Strategy;
+import org.apache.struts2.ognl.OgnlReflectionProvider;
 import org.apache.struts2.ognl.ProviderAllowlist;
+import org.apache.struts2.result.Result;
 import org.apache.struts2.result.ServletDispatcherResult;
+import org.apache.struts2.util.TextParseUtil;
+import org.apache.struts2.util.fs.DefaultFileManager;
+import org.apache.struts2.util.fs.DefaultFileManagerFactory;
+import org.apache.struts2.util.reflection.ReflectionException;
 import org.easymock.EasyMock;
 
-import jakarta.servlet.ServletContext;
 import java.net.MalformedURLException;
 import java.util.ArrayList;
 import java.util.Collections;
