+++ plugins/json/src/test/java/org/apache/struts2/json/JSONResultTest.java
@@ -18,19 +18,19 @@
  */
 package org.apache.struts2.json;
 
+import jakarta.servlet.http.HttpServletResponse;
 import org.apache.struts2.ActionContext;
-import org.apache.struts2.Result;
-import org.apache.struts2.mock.MockActionInvocation;
-import org.apache.struts2.util.ValueStack;
 import org.apache.struts2.StrutsStatics;
 import org.apache.struts2.junit.StrutsTestCase;
 import org.apache.struts2.junit.util.TestUtils;
+import org.apache.struts2.mock.MockActionInvocation;
+import org.apache.struts2.result.Result;
+import org.apache.struts2.util.ValueStack;
 import org.springframework.aop.framework.ProxyFactory;
 import org.springframework.mock.web.MockHttpServletRequest;
 import org.springframework.mock.web.MockHttpServletResponse;
 import org.springframework.mock.web.MockServletContext;
 
-import jakarta.servlet.http.HttpServletResponse;
 import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.text.SimpleDateFormat;
