+++ plugins/spring/src/test/java/org/apache/struts2/spring/SpringObjectFactoryTest.java
@@ -23,7 +23,6 @@ import org.apache.struts2.ActionChainResult;
 import org.apache.struts2.ActionContext;
 import org.apache.struts2.ModelDrivenAction;
 import org.apache.struts2.ObjectFactory;
-import org.apache.struts2.Result;
 import org.apache.struts2.SimpleAction;
 import org.apache.struts2.TestBean;
 import org.apache.struts2.TextProvider;
@@ -36,13 +35,14 @@ import org.apache.struts2.inject.ContainerBuilder;
 import org.apache.struts2.inject.Inject;
 import org.apache.struts2.interceptor.Interceptor;
 import org.apache.struts2.interceptor.ModelDrivenInterceptor;
+import org.apache.struts2.interceptor.NoOpInterceptor;
 import org.apache.struts2.mock.DummyTextProvider;
+import org.apache.struts2.result.Result;
 import org.apache.struts2.test.StubConfigurationProvider;
 import org.apache.struts2.util.location.LocatableProperties;
 import org.apache.struts2.validator.Validator;
 import org.apache.struts2.validator.validators.ExpressionValidator;
 import org.apache.struts2.validator.validators.RequiredStringValidator;
-import org.apache.struts2.interceptor.NoOpInterceptor;
 import org.springframework.aop.framework.Advised;
 import org.springframework.aop.framework.autoproxy.BeanNameAutoProxyCreator;
 import org.springframework.aop.interceptor.DebugInterceptor;
